Add support to downscale daemonsets
There might be some charts who use daemonsets, for example
prometheus-operator uses daemonsets to deploy node-exporter pods on each
node.
Since the way to scale them is different than scalable resources like
deployments, this needs to be addressed separately.
To scale a daemonset effectively means to disable them on all node by
using a `nodeSelector` replacing the current one, or adding if not
configured.
The nodeSelector value is `{'non-existing': 'true'}`, thus forcing the
daemonset pods to not be deployed on any nodes.
The full path of the used property is:
`spec.template.spec.nodeSelector.non-existing`
We also need to take care that when the scale up happens the original
nodeSelector of the daemonset is set back on the resource by keeping the
value in the annotation `downscaler/original-nodeSelector`.
There is no configuration option for the downtime nodeSelector, this being
hardcoded to `{'non-existing': 'true'}`.

The original autoscale_resource logic is untouched to ensure regression,
and the implemenation to autoscale daemonsets has been extracted to a
separete function autoscale_daemonsets which takes the first argument
as Daemonset instead of NamespaceAPIObject.

Added tests for autoscaling daemonsets.

Also updated the error message for --include-resources to provide an
example of the right way to pass the arguments for this flag.